fix possible failure cause by RLIMIT_MEMLOCK value too low
bpf_prog_load() may fail with EPERM because of a memory limit error.
We try to catchg this error by setting a value for `RLIMIT_MEMLOCK` tha we know enough
to load the ebpf program.
Note: If the ebpf program size increases than the rlimit values may need to be updated.