fedora: fix build
"uname -r" reports this for me:
4.8.10-200.fc24.x86_64

The x86_64 part was missing

Also, the distro name is now based on the ID field in /etc/os-release
(i.e. "fedora" instead of "fedora-24"). This will allow Scope to detect
the distribution.

The distro name has a specific Makefile rule "distro-id" in each
environments/*.mk. For now, this is the same parsing of /etc/os-release
but it will be different in CoreOS because CoreOS will not be build from
a coreos container but maybe a fedora container.