Guess family before sport
The source port appears twice in the kernel structure, one in host
endianness in (struct sock_common).skc_num and another one in network
endianness in (struct inet_sock)->inet_sport.

To avoid returning the first one if it happens that the source port used
is the same in both endiannesses, we start guessing at the family
offset.